Add files via upload
In this project, I tackled the challenge of designing a water billing system, gaining valuable experience as a Business Analyst. I collaborated with developers, creating product backlogs, use cases, and wireframes to ensure a user-friendly experience. My focus was on usability testing, identifying and reporting bugs through Github to improve functionality. This project honed my technical skills and solidified the importance of teamwork in agile development.